+++ CHANGELOG.md
@@ -42,6 +42,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * `jj fix` now buffers lines from subprocesses' stderr streams and emits them a
   complete line at a time. Each line is prepended with the file name.
 
+* `jj status` now collapses fully untracked directories into one line.
+  It still fully traverses them while snapshotting but they won't clutter up
+  the output with all of their contents.
+
 ### Fixed bugs
 
 ### Packaging changes